requirements.txt
disnake
python-dotenv
flask[async]
aiohttp
asyncio
psycopg2-binary
